refactor: 폴더/파일 구조 변경
