+++ types/Component.ts
